Restructured dir to python packaging standards
